Use docker to build packages Build deb w/ debhelper instead of electron-builder Remove ratpoison as OctoDash is the sole window. Create systemd service to launch OctoDash directly without a tty shell. Set systemd target to graphical. Incorperate xinit script into package. Create symlink for /usr/bin/octodash -> /opt/OctoDash/octodash Modify install/remove scripts to re-execute self with sudo. Deprecate using remove.sh. Removing via apt should be sufficient now. Add --version switch to octodash. Move release URL to environment variable Disable security when in dev mode (Bypass CORS checks) Remove update.sh Prevent remove.sh from running on newer installs

+++ scripts/remove.sh
@@ -1,16 +1,43 @@
 #!/bin/bash
 
+if [ "$EUID" -ne 0 ]
+    echo "This script requires root privileges. Please enter your sudo password when prompted."
+    sudo $0 $@
+    exit
+fi
+
+VERSION=$(dpkg -s octodash 2>/dev/null| grep '^Version' | awk ' { print $2 } ')
+
+if [ -z "${VERSION}" ]
+then
+    echo "Not installed"
+    exit 1
+fi
+
+VER_MAJOR=$(echo "${VERSION} | awk -F'.' ' { print $1 } '")
+VER_MINOR=$(echo "${VERSION} | awk -F'.' ' { print $2 } '")
+VER_PATCH=$(echo "${VERSION} | awk -F'.' ' { print $3 } '")
+
+if ![[ $VER_MAJOR -le 2 && $VER_MINOR -le 1 && $VER_PATCH -le 2 ]]
+then
+    echo "This script is intended for OctoDash 2.1.2 and earlier. To remove execute the following commands:"
+    echo -e "\t systemctl stop octodash"
+    echo -e "\t apt remove octodash"
+    exit 1
+fi
+
 echo "Removing OctoDash ..."
+
 killall octodash
 
-sudo dpkg -P octodash
+dpkg -P octodash
 
-rm -rf ~/.config/octodash/
+rm -rf /home/pi/.config/octodash/
 
-sed -i '/xset s off/d' ~/.xinitrc
-sed -i '/xset s noblank/d' ~/.xinitrc
-sed -i '/xset -dpms/d' ~/.xinitrc
-sed -i '/ratpoison&/d' ~/.xinitrc
-sed -i '/octodash/d' ~/.xinitrc
+sed -i '/xset s off/d' /home/pi/.xinitrc
+sed -i '/xset s noblank/d' /home/pi/.xinitrc
+sed -i '/xset -dpms/d' /home/pi/.xinitrc
+sed -i '/ratpoison&/d' /home/pi/.xinitrc
+sed -i '/octodash/d' /home/pi/.xinitrc
 
-echo "OctoDash has been removed :(. Please review your ~/.xinitrc and ~/.bashrc files to make sure everything got removed properly!"
+echo "OctoDash has been removed :(. Please review your /home/pi/.xinitrc and /home/pi/.bashrc files to make sure everything got removed properly!"
